+++ components/bottombar/chat/comment-item.tsx
