+++ components/bottombar/note.stories.tsx
