+++ components/bottombar/note.tsx
