+++ components/editor/index.tsx
