+++ components/editor/panel.tsx
