+++ components/sidebar/entry-list.tsx
